added trails and users services and rewrote trail update to comply
