+++ app/templates/root.tpl.html
@@ -1,2 +1,7 @@
 <h1>root</h1>
+<nav>
+	<span><a href="#/home">Home</a></span>
+	<span><a href="#/trails/new">New Trail</a></span>
+	<span><a href="#/topTrails">Top Trails</a></span>
+</nav>
 <div ui-view></div>
