+++ app/templates/topTrails.tpl.html
@@ -1,0 +1,1 @@
+<h1>Top Trails</h1>
