+++ src/js/controllers/root.js
@@ -1,19 +1,19 @@
-function RootController (OwnerService, $rootScope) {
+function RootController ($rootScope) {
 	let vm = this;
-	vm.admin = OwnerService.isAdmin();
-	vm.loggedIn = OwnerService.isLoggedIn();
-	vm.logout = logout;
+	// vm.admin = OwnerService.isAdmin();
+	// vm.loggedIn = OwnerService.isLoggedIn();
+	// vm.logout = logout;
 
-	$rootScope.$on('loginChange', (event, data) => {
-		vm.loggedIn = OwnerService.isLoggedIn();
-		vm.admin = OwnerService.isAdmin();
-	});
+	// $rootScope.$on('loginChange', (event, data) => {
+	// 	vm.loggedIn = OwnerService.isLoggedIn();
+	// 	vm.admin = OwnerService.isAdmin();
+	// });
 
-	function logout(){
-	  	OwnerService.logout()
-	    $rootScope.$broadcast('loginChange', {});
-  	}
+	// function logout(){
+	//   	OwnerService.logout()
+	//     $rootScope.$broadcast('loginChange', {});
+ //  	}
 }
 
- RootController.$inject = ['OwnerService', '$rootScope'];
+ RootController.$inject = ['$rootScope'];
  export {RootController};
