+++ src/js/controllers/topTrails.js
@@ -1,0 +1,6 @@
+function TopTrailsController () {
+
+}
+
+TopTrailsController.$inject = [];
+export {TopTrailsController}
