+++ src/js/main.js
@@ -6,12 +6,14 @@ import 'ngMap';
 
 import {RootController} from "./controllers/root";
 import {HomeController} from "./controllers/home";
+import {TopTrailsController} from "./controllers/topTrails"
 
 import {TrailController} from "./controllers/trail";
 import {TrailNewController} from "./controllers/trailNew";
 import {TrailUpdateController} from "./controllers/trailUpdate";
 
-import {OwnerService} from "./services/owner";
+import {UsersService} from "./services/users";
+import {TrailsService} from "./services/trails";
 
 
 
@@ -22,4 +24,7 @@ angular.module('app', ['ngMap', 'ui.router', 'ngCookies'])
 	.controller('TrailController', TrailController)
 	.controller('TrailNewController', TrailNewController)
 	.controller('TrailUpdateController', TrailUpdateController)
-	.service('OwnerService', OwnerService)
+	.controller('TopTrailsController', TopTrailsController)
+
+	.service('UsersService', UsersService)
+	.service('TrailsService', TrailsService)
